% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_location_shift.R
\name{ci_median_diff}
\alias{ci_median_diff}
\title{CI for the Population Median Difference of two Samples}
\usage{
ci_median_diff(
  x,
  y,
  probs = c(0.025, 0.975),
  type = "bootstrap",
  boot_type = c("bca", "perc", "norm", "basic"),
  R = 9999,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{probs}{Lower and upper probabilities, by default c(0.025, 0.975).}

\item{type}{Type of CI. Currently, "bootstrap" is the only option.}

\item{boot_type}{Type of bootstrap CI ("bca", "perc", "norm", "basic").}

\item{R}{The number of bootstrap resamples.}

\item{seed}{An integer random seed.}

\item{...}{Further arguments passed to \code{boot::boot()}.}
}
\value{
An object of class "cint" containing these components:
\itemize{
\item \code{parameter}: Parameter specification.
\item \code{interval}: CI for the parameter.
\item \code{estimate}: Parameter estimate.
\item \code{probs}: Lower and upper probabilities.
\item \code{type}: Type of interval.
\item \code{info}: Additional description.
}
}
\description{
This function calculates bootstrap CIs for the population value of
median(x) - median(y) by calling ci_quantile_diff(, q = 0.5).
See \code{\link{ci_quantile_diff}} for details.
}
\examples{
x <- 10:30
y <- 1:30
ci_median_diff(x, y, R = 999)  # Use larger value for R
}
\seealso{
\code{\link{ci_quantile_diff}}.
}
