% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_measures_of_scale.R
\name{ci_var}
\alias{ci_var}
\title{CI for the Population Variance}
\usage{
ci_var(
  x,
  probs = c(0.025, 0.975),
  type = c("chi-squared", "bootstrap"),
  boot_type = c("bca", "perc", "stud", "norm", "basic"),
  R = 9999L,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{probs}{Lower and upper probabilities, by default c(0.025, 0.975).}

\item{type}{Type of CI. One of \code{"chi-squared"} (default) or \code{"bootstrap"}.}

\item{boot_type}{Type of bootstrap CI. Only used for \code{type = "bootstrap"}.}

\item{R}{The number of bootstrap resamples. Only used for \code{type = "bootstrap"}.}

\item{seed}{An integer random seed. Only used for \code{type = "bootstrap"}.}

\item{...}{Further arguments passed to \code{\link[boot:boot]{boot::boot()}}.}
}
\value{
An object of class "cint", see \code{\link[=ci_mean]{ci_mean()}} for details.
}
\description{
This function calculates CIs for the population variance.
}
\details{
By default, classic CIs are calculated based on the chi-squared distribution,
assuming normal distribution (see Smithson). Bootstrap CIs are also available
(default: "bca"). We recommend them for the non-normal case.

The \code{stud} (bootstrap t) bootstrap uses the standard error of the sample variance
given in Wilks.
}
\examples{
x <- 1:100
ci_var(x)
ci_var(x, type = "bootstrap", R = 999)  # Use larger R
}
\references{
\enumerate{
\item Smithson, M. (2003). Confidence intervals. Series: Quantitative Applications in
the Social Sciences. New York, NY: Sage Publications.
\item S.S. Wilks (1962), Mathematical Statistics, Wiley & Sons.
}
}
\seealso{
\code{\link[=ci_sd]{ci_sd()}}
}
