% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramersv.R
\name{cramersv}
\alias{cramersv}
\title{Cramér's V}
\usage{
cramersv(x)
}
\arguments{
\item{x}{A \code{data.frame} with exactly two columns.}
}
\value{
A numeric vector of length one.
}
\description{
This function calculates Cramér's V, a measure of association between two categorical variables. It is a scaled version of the chi-squared test statistic and lies between 0 and 1. Cramér's V is calculated as sqrt(chi-squared / (n * (k - 1))), where n is the number of observations and k is the smaller of the number of levels of the two variables.
}
\details{
Note that no continuity correction is applied to the 2x2 case.
}
\examples{
ir <- iris
ir$PL <- ir$Petal.Width > 1
cramersv(ir[, c("Species", "PL")])
}
\references{
Cramer, Harald. 1946. Mathematical Methods of Statistics. Princeton: Princeton University Press, page 282 (Chapter 21. The two-dimensional case).
}
