% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{is.cint}
\alias{is.cint}
\title{Type Check}
\usage{
is.cint(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
A logical vector of length one.
}
\description{
Checks if an object inherits class "cint".
}
\examples{
is.cint(ci_proportion(5, 20))
is.cint(c(1, 2))
}
