% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_and_kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{Pearson's Measure of Kurtosis}
\usage{
kurtosis(z, na.rm = TRUE)
}
\arguments{
\item{na.rm}{Logical flag indicating whether to remove missing values or not.
Default is \code{TRUE}.}

\item{x}{A numeric vector.}
}
\value{
Numeric vector of length 1.
}
\description{
Defined as the ratio of the 4th central moment and the squared
second central moment. Under perfect normality, the kurtosis equals 3.
}
\examples{
kurtosis(1:10)
kurtosis(rnorm(1000))
}
\seealso{
\code{\link{moment}}
}
