% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{print.cint}
\alias{print.cint}
\title{Print "cint" Object}
\usage{
\method{print}{cint}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{A on object of class "cint".}

\item{digits}{Number of digits used to format numbers.}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for an object of class "cint".
}
\examples{
ci_mean(1:100)
}
