% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{se}
\alias{se}
\alias{se_mean}
\alias{se_var}
\alias{se_proportion}
\title{Standard errors}
\usage{
se_mean(z, na.rm = TRUE, ...)

se_var(z, na.rm = TRUE, ...)

se_proportion(z, na.rm = TRUE, ...)
}
\arguments{
\item{z}{Numeric vector.}

\item{na.rm}{Should missing values be removed before calculation? The default is \code{TRUE} for convenience.}

\item{...}{Further arguments to be passed from other methods.}
}
\value{
A numeric vector of length one.
}
\description{
Functions to calculate standard errors of different statistics. The availability of a standard error (or statistic proportional to it) allows to apply "bootstrapT" bootstrap.
}
\examples{
se_mean(1:100)
}
