% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_and_kurtosis.R
\name{skewness}
\alias{skewness}
\title{Sample Skewness}
\usage{
skewness(z, na.rm = TRUE)
}
\arguments{
\item{na.rm}{Logical flag indicating whether to remove missing values or not.
Default is \code{TRUE}.}

\item{x}{A numeric vector.}
}
\value{
Numeric vector of length 1.
}
\description{
Calculates sample skewness, i.e., the third central moment. A value of 0 refers to
a perfectly symmetric empirical distribution.
}
\examples{
skewness(1:10)
skewness(rexp(100))
}
\seealso{
\code{\link{moment}}
}
