% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boot.R
\name{ci_boot}
\alias{ci_boot}
\title{Bootstrap CI}
\usage{
ci_boot(S, boot_type = c("norm", "basic", "stud", "perc", "bca"), probs)
}
\arguments{
\item{S}{Result of \code{boot::boot()}.}

\item{boot_type}{CI type passed to \code{boot::boot.ci()}.}

\item{probs}{Lower and upper probabilities.}
}
\value{
A numeric vector of length two.
}
\description{
Internal wrapper to \code{boot::boot.ci()}.
Contains the logic for also calculating one-sided CIs, and for unequal-tailed CIs.
}
