% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_boot.R
\name{ci_boot}
\alias{ci_boot}
\title{Bootstrap Confidence Interval}
\usage{
ci_boot(S, boot_type = c("norm", "basic", "stud", "perc", "bca"), probs, ...)
}
\arguments{
\item{S}{Result of \code{boot}.}

\item{boot_type}{Confidence interval type passed to \code{boot::boot.ci}.}

\item{probs}{Probability cuts for the confidence intervals.}

\item{...}{Further arguments passed to \code{boot::boot.ci}.}
}
\value{
A numeric vector of length two.
}
\description{
Internel wrapper to \code{boot::boot.ci}. Contains the logic for one-sided and asymmetric confidence intervals.
}
