% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_cor.R
\name{ci_cor}
\alias{ci_cor}
\title{CI for Correlation Coefficients}
\usage{
ci_cor(
  x,
  y = NULL,
  probs = c(0.025, 0.975),
  method = c("pearson", "kendall", "spearman"),
  type = c("normal", "bootstrap"),
  boot_type = c("bca", "perc", "norm", "basic"),
  R = 9999,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector or a \code{matrix/data.frame} with exactly two numeric columns.}

\item{y}{A numeric vector (only used if \code{x} is a vector).}

\item{probs}{Probabilites. The default c(0.025, 0.975) gives a symmetric 95\% CI.}

\item{method}{Type of correlation coefficient, one of "pearson" (default), "kendall",
or "spearman". For the latter two, only bootstrap CIs are supported.
The names can be abbreviated.}

\item{type}{Type of CI. One of "normal" (the default) or "bootstrap"
(the only option for rank-correlations).}

\item{boot_type}{Type of bootstrap CI ("bca", "perc", "norm", "basic").
Only used for \code{type = "bootstrap"}.}

\item{R}{The number of bootstrap resamples. Only used for \code{type = "bootstrap"}.}

\item{seed}{An integer random seed. Only used for \code{type = "bootstrap"}.}

\item{...}{Further arguments passed to \code{boot::boot()}.}
}
\value{
An object of class "cint" containing these components:
\itemize{
\item \code{parameter}: The parameter in question.
\item \code{interval}: The CI for the parameter.
\item \code{estimate}: The estimate for the parameter.
\item \code{probs}: A vector of error probabilities.
\item \code{type}: The type of the interval.
\item \code{info}: An additional description text for the interval.
}
}
\description{
This function calculates CIs for a population correlation coefficient.
For Pearson correlation, "normal" CIs are available (by \code{stats::cor.test()}).
Also bootstrap CIs are supported (by default "bca", and the only option for rank correlations).
}
\examples{
ci_cor(iris[1:2])
ci_cor(iris[1:2], type = "bootstrap", R = 999)  # Use larger R
ci_cor(iris[1:2], method = "spearman", type = "bootstrap", R = 999)  # Use larger R
}
