% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_cramersv.R
\name{ci_cramersv}
\alias{ci_cramersv}
\title{CI for the Population Cramer's V}
\usage{
ci_cramersv(
  x,
  probs = c(0.025, 0.975),
  type = c("chi-squared", "bootstrap"),
  boot_type = c("bca", "perc", "norm", "basic"),
  R = 9999,
  seed = NULL,
  test_adjustment = TRUE,
  ...
)
}
\arguments{
\item{x}{The result of \code{stats::chisq.test()}, a matrix/table of counts,
or a \code{data.frame} with exactly two columns representing the two variables.}

\item{probs}{Probabilites. The default c(0.025, 0.975) gives a symmetric 95\% CI.}

\item{type}{Type of CI. One of "chi-squared" (default) or "bootstrap".}

\item{boot_type}{Type of bootstrap CI ("bca", "perc", "norm", "basic").
Only used for \code{type = "bootstrap"}.}

\item{R}{The number of bootstrap resamples. Only used for \code{type = "bootstrap"}.}

\item{seed}{An integer random seed. Only used for \code{type = "bootstrap"}.}

\item{test_adjustment}{Adjustment to allow for test of association, see Details.
The default is \code{TRUE}.}

\item{...}{Further arguments passed to \code{boot::boot()}.}
}
\value{
An object of class "cint" containing these components:
\itemize{
\item \code{parameter}: The parameter in question.
\item \code{interval}: The CI for the parameter.
\item \code{estimate}: The estimate for the parameter.
\item \code{probs}: A vector of error probabilities.
\item \code{type}: The type of the interval.
\item \code{info}: An additional description text for the interval.
}
}
\description{
This function calculates CIs for the population Cramer's V.
By default, a parametric approach based on the non-centrality parameter (NCP)
of the chi-squared distribution is utilized.
Alternatively, bootstrap CIs are available (by default "bca"),
also by boostrapping CIs for the NCP.
}
\details{
A positive lower (1-alpha)*100\%-confidence limit for the NCP goes hand-in-hand with a
significant association test at level alpha. In order to allow such test approach
also with Cramer's V, if the lower bound for the NCP is 0,
we round down to 0 also the lower bound for Cramer's V.
Without this slightly conservative adjustment, the lower limit for V would always be
positive since ci for V = sqrt((ci for NCP + df)/(n (k - 1))), where k is the smaller
number of levels in the two variables (see Smithson for this formula).
Use \code{test_adjustment = FALSE} to switch off this behaviour. Note that this is
also a reason to bootstrap V via NCP instead of directly bootstrapping V.

Further note that no continuity correction is applied for 2x2 tables,
and that large chi-squared test statistics might provide unreliable results with
method "chi-squared" (see \code{?pchisq}).
}
\examples{
ci_cramersv(mtcars[c("am", "vs")])
}
\references{
Smithson, M. (2003). Confidence intervals. Series: Quantitative Applications in the Social Sciences. New York, NY: Sage Publications.
}
\seealso{
\code{\link{ci_chisq_ncp}}.
}
