% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_median.R
\name{ci_median}
\alias{ci_median}
\title{CI for the Population Median}
\usage{
ci_median(
  x,
  probs = c(0.025, 0.975),
  type = c("binomial", "bootstrap"),
  boot_type = c("bca", "perc", "norm", "basic"),
  R = 9999,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{probs}{Lower and upper probabilities, by default c(0.025, 0.975).}

\item{type}{Type of CI. One of "binomial" (default), or "bootstrap".}

\item{boot_type}{Type of bootstrap CI ("bca", "perc", "norm", "basic").
Only used for \code{type = "bootstrap"}.}

\item{R}{The number of bootstrap resamples. Only used for \code{type = "bootstrap"}.}

\item{seed}{An integer random seed. Only used for \code{type = "bootstrap"}.}

\item{...}{Further arguments passed to \code{boot::boot()}.}
}
\value{
An object of class "cint" containing these components:
\itemize{
\item \code{parameter}: Parameter specification.
\item \code{interval}: CI for the parameter.
\item \code{estimate}: Parameter estimate.
\item \code{probs}: Lower and upper probabilities.
\item \code{type}: Type of interval.
\item \code{info}: Additional description.
}
}
\description{
This function calculates CIs for the population median by calling
\code{ci_quantile(..., q = 0.5)}. See \code{\link{ci_quantile}} for details.
}
\examples{
ci_median(1:100)
}
\seealso{
\code{\link{ci_quantile}}.
}
