% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_proportion.R
\name{ci_proportion}
\alias{ci_proportion}
\title{CI for a Population Proportion}
\usage{
ci_proportion(
  x,
  n = NULL,
  probs = c(0.025, 0.975),
  type = c("Clopper-Pearson", "Agresti-Coull", "Wilson", "bootstrap"),
  boot_type = c("bca", "perc", "stud", "norm", "basic"),
  R = 9999,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector of 0 and 1 or the number of successes.}

\item{n}{The sample size. Only needed if \code{x} is a vector of length 1.}

\item{probs}{Probabilites. The default c(0.025, 0.975) gives a symmetric 95\% CI.}

\item{type}{Type of CI. One of "Clopper-Pearson" (the default), "Agresti–Coull",
"Wilson", "bootstrap".}

\item{boot_type}{Type of bootstrap CI ("bca", "perc", "stud", "norm", "basic").
Only used for \code{type = "bootstrap"}.}

\item{R}{The number of bootstrap resamples. Only used for \code{type = "bootstrap"}.}

\item{seed}{An integer random seed. Only used for \code{type = "bootstrap"}.}

\item{...}{Further arguments passed to \code{boot::boot()}.}
}
\value{
An object of class "cint" containing these components:
\itemize{
\item \code{parameter}: The parameter in question.
\item \code{interval}: The CI for the parameter.
\item \code{estimate}: The estimate for the parameter.
\item \code{probs}: A vector of error probabilities.
\item \code{type}: The type of the interval.
\item \code{info}: An additional description text for the interval.
}
}
\description{
This function calculates CIs for a population proportion. By default,
"Clopper-Pearson" CIs are calculated (via \code{stats::binom.test()}).
Further possibilities are "Wilson", "Agresti-Coull", and "bootstrap"
(not recommended; by default "bca").
}
\examples{
x <- rep(0:1, times = c(50, 100))
ci_proportion(x)
ci_proportion(x, type = "Wilson")
ci_proportion(x, type = "Agresti-Coull")
}
\references{
\enumerate{
\item Clopper, C. and Pearson, E. S. (1934). The use of confidence or fiducial limits illustrated in the case of the binomial. Biometrika. 26 (4).
\item Wilson, E. B. (1927). Probable inference, the law of succession, and statistical inference. Journal of the American Statistical Association, 22 (158).
\item Agresti, A. and Coull, B. A. (1998). Approximate is better than 'exact' for interval estimation of binomial proportions. The American Statistician, 52 (2).
}
}
