% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_skew_and_kurtosis.R
\name{ci_skewness}
\alias{ci_skewness}
\title{CI for the Skewness}
\usage{
ci_skewness(
  x,
  probs = c(0.025, 0.975),
  type = "bootstrap",
  boot_type = c("bca", "perc", "norm", "basic"),
  R = 9999,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{probs}{Probabilites. The default c(0.025, 0.975) gives a symmetric 95\% CI.}

\item{type}{Type of CI. Currently not used as the only type is \code{"bootstrap"}.}

\item{boot_type}{Type of bootstrap CI ("bca", "perc", "norm", "basic").}

\item{R}{The number of bootstrap resamples.}

\item{seed}{An integer random seed.}

\item{...}{Further arguments passed to \code{boot::boot()}.}
}
\value{
An object of class "cint" containing these components:
\itemize{
\item \code{parameter}: The parameter in question.
\item \code{interval}: The CI for the parameter.
\item \code{estimate}: The estimate for the parameter.
\item \code{probs}: A vector of error probabilities.
\item \code{type}: The type of the interval.
\item \code{info}: An additional description text for the interval.
}
}
\description{
This function calculates bootstrap CIs for the population skewness.
By default, bootstrap type "bca" is used.
}
\examples{
x <- 1:20
ci_skewness(x, R = 999)  # In practice, use larger R
}
\seealso{
\code{\link{kurtosis}}
}
