% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_skew_and_kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{Pearson's Measure of Kurtosis}
\usage{
kurtosis(z, na.rm = TRUE)
}
\arguments{
\item{z}{A numeric vector.}

\item{na.rm}{Logical flag indicating whether to remove missing values or not.
Default is \code{TRUE}.}
}
\value{
Numeric vector of length 1.
}
\description{
Defined as the ratio of the 4th central moment and the squared
second central moment. Under perfect normality, the kurtosis equals 3.
Put differently, we do not show "excess kurtosis" but rather kurtosis.
}
\examples{
kurtosis(1:10)
kurtosis(rnorm(1000))
}
\seealso{
\code{\link[=moment]{moment()}}, \code{\link[=skewness]{skewness()}}
}
