% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cint.R
\name{print.cint}
\alias{print.cint}
\title{Print a cint Object}
\usage{
\method{print}{cint}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{A on object of class \code{cint}.}

\item{digits}{Number of digits used to format sample estimate and confidence limits.}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for an object of class \code{cint}.
}
\examples{
ci_mean(1:100)
}
