% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_skew_and_kurtosis.R
\name{skewness}
\alias{skewness}
\title{Sample Skewness}
\usage{
skewness(z, na.rm = TRUE)
}
\arguments{
\item{z}{A numeric vector.}

\item{na.rm}{Logical flag indicating whether to remove missing values or not.
Default is \code{TRUE}.}
}
\value{
Numeric vector of length 1.
}
\description{
Calculates sample skewness. A value of 0 refers to a perfectly symmetric distribution.
}
\examples{
skewness(1:10)
skewness(rexp(100))
}
\seealso{
\code{\link[=moment]{moment()}}, \code{\link[=kurtosis]{kurtosis()}}
}
